Fixed #55457(cli built-in web server does not work with IE), introduced by r315128

+++ sapi/cli/php_http_parser.h
@@ -112,7 +112,7 @@ struct php_http_parser {
   unsigned char index;
 
   uint32_t nread;
-  size_t content_length;
+  long content_length;
 
   /** READ-ONLY **/
   unsigned short http_major;
